+++ specifiers.cpp
@@ -28,6 +28,7 @@
 #include <stdlib.h>
 
 #include "sscanf.h"
+#include "args.h"
 #include "utils.h"
 #include "data.h"
 #include "specifiers.h"
