+++ sscanf.vcxproj.filters
@@ -27,6 +27,7 @@
     <ClInclude Include="specifiers.h" />
     <ClInclude Include="sscanf.h" />
     <ClInclude Include="utils.h" />
+    <ClInclude Include="args.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="SDK\amxplugin.cpp">
